requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cloudinary==1.24.0
confusable-homoglyphs==3.2.0
Django==1.11.29
django-bootstrap3==14.2.0
django-registration==2.4.1
django-tinymce==3.2.0
importlib-metadata==1.7.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
